pyproject.toml
[project]
name = "demo-github-action-predictions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "nixtla>=0.6.6",
    "pandas-stubs>=2.2.3.241126",
    "yfinance>=0.2.54",
]
